requirements.txt
requests
beautifulsoup4
slackclient
